+++ server.js
@@ -7,16 +7,20 @@ const app = express();
 
 // Endpoints folders
 const children = require("./routes/childrenRoutes");
-const tasks = require("./routes/tasksRoutes");
 const icons = require("./routes/iconsRoutes");
+const tasks = require("./routes/tasksRoutes");
+const goals = require("./routes/childGoalsRoutes");
+const childTasks = require("./routes/childTasksRoutes");
 
 // Server functionalities
 app.use(cors());
 app.use(express.json());
 app.use(express.static("./assets"));
 app.use("/children", children);
-app.use("/tasks", tasks);
 app.use("/icons", icons);
+app.use("/tasks", tasks);
+app.use("/children", goals);
+app.use("/children", childTasks);
 
 // Server launching
 app.listen(PORT, () => {
